NOT WORKING VERSION. Added the 0 max_no_of_gas case. The swingby periapsis and incoming velocity parameter is added to optimisation

+++ mga-ltto/mgso_optimisation.py
@@ -87,28 +87,29 @@ if __name__ == '__main__': #to prevent this code from running if this file is no
     """
     
     # mgso General parameters
-    max_no_of_gas = 2
-    no_of_sequence_recursions = max_no_of_gas # amount of different predefiend islands
+    max_no_of_gas = 0
+    # no_of_sequence_recursions = 1 if max_no_of_gas == 0 else max_no_of_gas # amount of different predefiend islands
+    no_of_sequence_recursions = max_no_of_gas
     max_number_of_exchange_generations = 1 # amount of times it evolves
-    # number_of_sequences_per_planet = 4
-    number_of_sequences_per_planet = [3 for _ in range(max_no_of_gas)]
-    # number_of_sequences_per_planet = [1 for i in range(max_no_of_gas)]
+    # number_of_sequences_per_planet = [1]
+    number_of_sequences_per_planet = [1 for _ in range(max_no_of_gas)]
     
     ## Specific parameters
     departure_planet = "Earth"
-    arrival_planet = "Jupiter"
+    arrival_planet = "Mars"
     free_param_count = 2
-    num_gen = 20
-    pop_size = 500
+    num_gen = 2
+    pop_size = 100
     assert pop_size > 62
     no_of_points = 1000
-    bounds = [[1000, 0, 50, -10**4, 0],
-            [1200, 0, 2000, 10**4, 6]]
-    # print('Departure date bounds : [%d, %d]' %
-    #         (time_conversion.julian_day_to_calendar_date(1000),
-    #     time_conversion.julian_day_to_calendar_date(1200)))
+    bounds = [[8000, 0, 200, 0, 2e2, -10**4, 0],
+            [8200, 0, 1200, 7000, 2e11, 10**4, 0]]
+    print('Departure date bounds : [%s, %s]' %
+            (time_conversion.julian_day_to_calendar_date(time_conversion.modified_julian_day_to_julian_day(bounds[0][0] + 51544.5)),
+        time_conversion.julian_day_to_calendar_date(time_conversion.modified_julian_day_to_julian_day(bounds[1][0]) + 51544.5)))
     subdirectory = '/mgso_full_test'
-    shutil.rmtree(output_directory + subdirectory)
+    if os.path.exists(output_directory + subdirectory):
+        shutil.rmtree(output_directory + subdirectory)
     # subdirectory = ''
 
     # num_gen = 1
